fix cast
implicit conversion from ‘float’ to ‘double’ when passing argument to function

+++ src/utils.c
@@ -835,7 +835,7 @@ int snprintf_pointer(char *dest, int max_len, int type, void *p,
 		break;
 
 	case VAR_FLOAT:
-		nb = snprintf(dest, max_len, "%f", (*(float *)p) * multiplier);
+		nb = snprintf(dest, max_len, "%f", (double)((*(float *)p) * multiplier));
 		break;
 
 	case VAR_HEX:
